requirements.txt
kafka-python==2.0.2
faker==19.13.0
flask==2.3.3
python-dotenv==1.0.0 